refactor(session-store): extract canonical config sanitization (file #4)
Lift config/capability sanitization + credential-redaction logic into
canonical-config-sanitize.ts (security-relevant logic now isolated + testable).
Pure transforms. check 0, store vitest parity (503 pass).
